+++ app/scripts/controller-init/utils.ts
@@ -154,10 +154,10 @@ export function initControllers({
     const persistedStateKey =
       persistedStateKeyRaw === null
         ? undefined
-        : persistedStateKeyRaw ?? controllerName;
+        : (persistedStateKeyRaw ?? controllerName);
 
     const memStateKey =
-      memStateKeyRaw === null ? undefined : memStateKeyRaw ?? controllerName;
+      memStateKeyRaw === null ? undefined : (memStateKeyRaw ?? controllerName);
 
     // @ts-expect-error: Union too complex.
     partialControllersByName[controllerName] = controller;